requirements.txt
BeautifulSoup==3.2.0
Django==1.5.5
Pillow==2.3.1
South==0.8.4
django-disqus==0.4.1
django-haystack==1.2.7
django-pagination==1.0.7
django-tagging==0.3.1
django-mobile==0.2.1
django-pipeline==1.3.21
django-pipeline-compass-rubygem==0.1.8
httplib2
oauth2==1.5.170
python-memcached==1.47
python-twitter==0.8.2
simplejson==2.1.6
smartypants==1.8.3
typogrify==2.0.2
xapian-haystack==1.1.5beta
akismet==0.2.0
Markdown==2.0.3
docutils
psycopg2==2.0.13
git+https://github.com/lazybird/django-bento.git
django-solo==1.0.4
django-imagekit==3.2.0
django-selectable==0.8.0

# tests
factory_boy
django-discover-runner

git+https://github.com/tylerball/django-oembed.git
git+https://github.com/tylerball/django-shorturls.git

# deployment
gunicorn
